Overhaul file placement to make dependencies easier to manage.
We really just split things in three files instead of two, but the focus
is now on functionality rather than OS support. OS-specific code is now
included inside the functionality-based class that makes sense for it.

p.s.: this layout is the one recommended by puppetlabs so litterature
can be found to justify it's usage and the anchor pattern.