Integer group values are not permitted by concat 1.2.x
The 1.2.x branch of the concat module is calling validate_string() from
stdlib to restrict value type on the $group argument. With the future
parser enabled, or with puppet 4.x, this means that the value of 0 that
we're giving to the concat resource is rejected.

This is unfortunate since integer values are premitted for file
resources.

See issue mentioned in added comment. The concat module should allow
integer values as well, but it's currently not the case.

[edit by: gabster@lelutin.ca] -- added the comment about the bug report
on the concat module and changed the commit message to make it more
descriptive.