Finish effort to parametrize findtime
Making the global value of findtime modifiable via a class parameter was
a good change. However, it was only implemented for RHEL. This means
that using this class parameter with debian machines doesn't do what the
user would expect it too.

So, let's make this parameter affect all supported Linux distributions.