Start purging jail.d by default
with the new parameter purge_jail_dot_d this purging can be disabled if
set to false.

in centos/redhat 6/7 and since debian jessie, fail2ban has a jail.d
directory. this dir can contain garbage files or duplicate definitions.
by purging, we ensure that we have a consistent configuration.

also, it'd be great to use that jail.d directory in the future to manage
jails in individual files in there instead of using concat fragments. in
that sense, purging the directory by default will mimic the current
behaviour with concat fragments: jails that are not managed through
puppet automatically get removed.