fileter includes: check that arrays are not empty
the "if @arrayname" lines seem to be verifying for the variable's
existence. this means that we're currently always getting an 'includes'
section with both 'before' and 'after' lines in filter config files.

with the .empty? method, we can make sure that we only output the
section if needed for at least one of the include lines, and that a line for
each element is present only if there is at least one element in the
corresponding array.