filter: missing ordering makes filter creation error out
During some runs, filter creation might trigger an error because of a
"file or directory not found". This is caused by our lack of ordering
with regards to the filter file: puppet might decide to try and create
that file before fail2ban is even installed.

To remediate this situation, we'll make filter files depend on the
'::config' class that is sure to be ordered after package installation.