iptables-multiport: define missing variables for debian
We're now managing contents of the iptables-multiport.conf file and
there are variables that exist in RHEL-based systems but not on debian.
This breaks the action entirely.

In order to accomodate RHEL overriding possibilities but still have
something that works in debian, let's just define those variables to
what's usually hardcoded in debian.