Don't keep jail.local around by default
Since the module used to force the contents of jail.local to setup
jails and these have either been moved to jail.d or (only for wheezy) to
jail.conf, then jail.local is superfluous.

Also, it ensure that configuration is in a consistent state.

Users can disable this by setting rm_jail_local to false.