remove gentoo code paths
Gentoo is currently not supported by this module. I can't verify that
the seemingly incomplete code paths that related to it were even
working. This code is dead so we should remove it.

If ppl want to add official support for gentoo (and help me to
continually test it), then I'd be glad to add the needed parts back in.