install missing file iptables-blocktype.conf on wheezy
originally on wheezy that file is not there, but if we just add it then
our current template for iptables-multiport works fine.

This fix means that ever since we've been using the template for
iptables-multiport, wheezy has been broken.

This change also fixes the before_include for debian stretch, which
should look similar to what redhat/centos receive.