only add missing variables on debian releases before stretch
since stretch, which uses fail2ban 0.9.x, the previously missing
variables in the iptables-multiport action definition are no longer
missing (they are in iptables-common.conf included at the top of the
action file).
So we want to stop including this hack starting with debian stretch.

With this, support for stretch is minimally in place. I haven't tested
all settings and resources yet but it's at least possible to install
fail2ban and create jails and have a working service.