Use built-in data types instead of validate_* functions from stdlib
This helps to shorten the code and make it more explicit what's expected
when looking at the parameter list.

Also, validate_* functions are deprecated. The validate_legacy is an
alternative to those functions, but instead of using that as an
intermediate step, we can just go ahead and use the data types. After
all, we've split puppet 3.x and 4.x support.