Replace pre-defined jail wrappers with data-in-module
This has the advantage that now there's only one defined type that needs
to be maintained where we can add parameters, not 25 of them.

It also makes it easier (and possible!) to override or set any parameter.

I've chosen to support hiera 4 which might be disputed by some ppl. This
choice is based on the fact that debian's package is shipping puppet 4.8
which still doesn't have hiera 5.

This commit also has some collateral damage changes since I was working
on this for way too long and it made it really hard to split things up:

 * parametrize the global default port parameter
 * make port and filter optional on the jail defined type since they
   have a global value that can be used.
 * move out some type definitions to custom types. this will make it
   easier to maintain the code in the long run.