README: inconsistencies in requirements and missing markup
The most significant change to the readme is the bit about which version
of stdlib is required by this module. The metadata.json file says
something different, so the README file should be corrected.

otherwise the rest is mostly about adding markup where class names or
file names are mentioned.