bump version to 3.1.0
this version brings a bunch of new paremeters: it's possible to
configure more of the fail2ban options in .conf files.

there is also now a deprecation notice that was added with regards to
persistent_bans, so I've decided to mark this as a minor version
upgrade. The api shouldn't become unstable for current users but it
might become annoying to folks who are using the persistent_bans param.