predefined;asterisk+freeswitch: drop email action
the default actions for most jails were copied over from examples that
existed in files from the debian package (they were probably taken from
upstream as well).

the asterisk and freeswitch jails send out an email to %(destemail)s by
default, which can be pretty annoying if this is not the behaviour that
you want. Let's drop the email action. Folks that want to have it can
easily override the value for the "action".