define $config_file_mode on jail and filter defines
The current implementation of config_file_mode doesn't work in jail and
filter: the variable with the same name in the main class is
inaccessible here. So let's just define the same parameter on the
defined types.

This means that setting config_file_mode on the main class will not
change permissions on jail files and filter files.