constrain debian jessie hack to the debian os family closes#43
xhejtman reported on github that with Centos and RedHat 8 being now
released, this condition for including the debian jessie hack for log
levels is being wrongly used.

We'll simply constrain to the debian os family. I think it should be
enough since ubuntu is way ahead in version numbers already.