README: mark intent to remove support for the 2.x branch
puppet 3.x is reallllly not relevant anymore. I've kept this branch
around for long enough and never got any changes proposed for it. I
personally also don't use this anymore so I'm not in a good place to
ensure support. the 2.x branch will still be there for folks who need it
but it won't get any merging going on for fixing any bugs.