Reorganize README and drop technical details
This change makes the readme a bit easier to read through. The structure
is a tiny bit more like puppetlabs' modules and there's a table of
contents to make HTML renderings easier to peruse.

I've also removed lists of parameters since they were out of date, and
the REFERENCE.md file (or an HTML redering of the puppet-strings
documentation) serves this purpose better.