README: add an example for how to create an action
An `action` defined type was recently added. Let's make it visible in
the README file so that users can better know that this type of
configuration is also possible to define with the module.

Let's bump nftables support documentation to being a subsection of the
actions section. It's now about using the `fail2ban::action` type per se
but it's still about configuring actions.