bump version to 2.0.3
Some fixes for older debian and redhat releases were sent in but not
released yet.

This marks the last release on the 2.x branch. Users can continue using
version 2.0.3 if they need it, but are strongly advised to upgrade to a
more recent version of puppet since 3.x has bee EOL for quite a while.