Match data type for logpath on main class
Since the parameter on the main class represents exactly the same data
as the parameter on the jail defined type but for the global
configuration, we should also change the data type for that global
setting.

Let's also change all default values and examples to arrays of strings.
This will prepare the way to eventually deprecate plain strings. The
deprecation will be added to the next minor release and the String type
will get removed in the next major release.