Implement the mode option for individual jails
Now that we have a global option for setting the mode for filters, it
would be nicer to offer it for individual jails, too. After all,
changing the "mode" for filters is way more relevant to do on individual
jails since we're only dealing with one filter at a time.