Change logpath's data type to only array of strings
I've written in the commit that was adding the array of strings type
that there would be a deprecation warning, but it's been a long while
and the code changes are currently heading towards a major release.

So I'm not being nice to users.. woops. But I'll add an upgrade note in
the readme about the change at least.

The reason for the change is that one can add multiple paths to the
logpath parameter, so it should be an array. Permitting only one data
type simplifies the code of the module.