Migrate all templates to epp syntax
In this module, only the template used in fail2ban::action was written
in epp syntax.

The templates are super simple in this module, but still it's a huge
gain to standardize on only one syntax.

I'd rather standardize on embedded puppet since it means one doesn't
have to know any ruby at all to interact with or comprehend this module.

Plus, explicitly passing the variables to the template is way better
than to have them magically appear in the template's runtime context by
"bleeding into" it.