Allow multiple actions to be configured through a list
According to fail2ban's documentation, action can take on multiple
values.

Currently, the only way to configure fail2ban this way with the module
is to hack our way by adding a newline and more actions in the one
string argument.

Now we can use an array of strings to only specify the action names and
their options.

The default value for the parameters are updated to a list of strings in
preparation for the eventual removal of the String type.

Using only a string is now considered as deprecated and it will get
removed in version 5.x of this module.