Teach all missing parameters to the jail defined type
All of the recently added parameters to the main class for configuring
global options in jail.conf can also be used in individual jails.

Let's also use the more precise types that were set in place for a
handful of parameters two commits back.

This changes the template for jails to be a simple iteration over
options. It simplifies maintenance of the template. The options that are
given as parameters to the defined type should always override the ones
in additional_options since they provide more precise type checking.

logpath, action and ignoreip need special formatting for additional
lines when more than one value is present. It would make more sense for
the template to add the additional formatting but it would require some
contortions to separate out the three values and would just be
confusing.