jail: the join()'ed values are always present in config but shouldn't be
When I changed things to format the values outside of the template, I
did that to simplify code in the template.

However, it introduced a bug where it was changing the type of those
values to always be a string, instead of being Undef when nothing was
specified in the params.

This implied that those values would always be output in jail config
files, possibly overriding the global values.

To fix that, I've added a snippet of code that automatically calculates
an offset for array values after the first one. This way we don't need
to manually format each option and we can let the template handle
things.

I've also tested out if ignoreip was being accepted when values are
split out into separate lines instead of just with a space, as the
jail(5) man page indicates: fail2ban correctly finds the values in this
case. So we don't need a special formatting for that option either.