Teach filter the missing parameters it can use
Most of those options could previously be configured with the
additional_defs array hack, but having them present in the code makes it
possible to document what they do.

additional_defs is still useful since users can define arbitrary
variables that'll be used in other options.

The init section is new however and was not possible to configure with
additional_defs since it would have placed failregex and ignoreregex in
the Init section.