make Debian 11 bullseye the default Debian configuration
This is for future-proofing. When Bookworm will come out, this module
would otherwise break again, because there won't be a `Debian-12.yaml`
file, and then it will revert to the defaults, which will not match
Debian's as well as this file does.

The idea here is that we keep Debian.yaml matching the latest stable,
then we make a copy in Debian-X.yaml *if* (and only if!) release X
differs from the latest stable.

Ideally, this would also allow us to get rid of *any* Debian-X.yaml
file once we stop supporting those old releases.