Use puppet syntax rake tasks for tests
Strangely I had already added a configuration option for this but I had
not included the tasks at the top of the file. I had not included the
gem in the Gemfile as a test dependency either.

I'll also try validating the hiera keys to see if that works well.