Update REFERENCE file with the correct version of puppet-strings
There was apparently a difference in the output between the version of
puppet-strings in debian and the one installed by the Gemfile. This
caused discrepancies that were always flagged as "the file being
outdated" in the CI checks.

Consequently, update the README so that I (and maybe others in the
future) don't forget how to correctly generate the file.