Disable puppet-syntax's hiera key check and reuse existing tasks
I tried earlier to enable puppet-syntax's hiera key checking. In normal
cases this would be super useful, but in the case of the fail2ban module
all of the defined keys are not meant to be used with puppet's autoload
so the check is just creating loads of useless errors. We could fix
those errors by changing the key names and make them respect puppet
autoload key syntax but that would be an annoying breaking change for
users. So for now, we won't use the hiera key check at all.

Both the custom :tests and :docs tasks are useless: puppet-syntax
defines a :validate task that runs all the same checks and
puppet-strings defines :strings:generate and :strings:generate:reference
that can handle everything for us.