Run puppet-rspec tests against each matrix combination
Sanity checks are already handled via the setup_matrix step. In the step
that uses the matrix, what we want is to run rspec tests against the
corresponding puppet version.

This will presently fail because there are still some elements missing
for the spec tests to run properly with what gets installed by the
Gemfile. I'll fix that in an upcoming commit.