Let the github action handle the puppet gem version spec
The leading "= " is problematic since the action is also injecting a
"~>" at the beginning of the version spec. Let's just let the action
handle the version specification format.

Users of the gemfile will also be able to specify the whole ruby gem
sepcification by themselves if they use the environment value.