Bump up all supported versions (puppet and OS) and drop EOL versions
Puppet 4 and 5 are currently EOL and running the CI tests against those
versions does not work. So support for those older versions is now
dropped from this module - meaning the code will not get tested against
those versions and fixes specifically for those older releases will only
get accepted if they are not adding useless (for supported puppet
versions) code branches

We're super late with supporting more recent linux releases! I'm
tentatively declaring all of the newer releases as supported for the
distros that are already marked. I haven't tested any of them but if
there's any issue that's reported we'll be able to fix them.

RHEL 6 is no longer supported upstream

CentOS 6 and 8 are EOL. Release 7 is still supported until 2024. I'm
presuming that supporting CentOS 8 comes with no additional cost so I'll
keep it around until we drop support for 7.

Debian buster is EOL but it's currently still in LTS.