Refresh the README text a bit
I don't use jenkins-ci anymore, but rather github actions. You should be
able to see the status of pipelines everywhere in the repository now.

The rest of the changes is mostly changes in wording in order to make
the text clearer and a bit easier to read.

I've also added some more details about how to customize where your
chain is placed if you use nftables.