Run tests against the most recent version of stdlib that we can
I've tried running tests against stdlib 9.0.0 to see if we could push
the requirement for stdlib up to that point and there was a deprecation
warning for the ensure_packages() function that was causing spec tests
to fail.

Fixing this deprecation warning would require us to call the function
with a different name, stdlib::ensure_packages(), but that name does not
exist in previous stdlib versions. So I don't want to make the change
just yet: I don't want to break compatibility with versions of stdlib
prior to 9.0.0.