Fix filter+mode for ssh and postfix jails (Closes: #74)
The jails for sshd-ddos and postfix-* redefine the `filter` argument so
that they use the one with the correct name. But since the new filter
name does not supply the `mode=...` parameter to the filter, this one is
not being used even though there's a `mode` variable, and this means
that the jails are not testing the correct thing! They're testing the
default "mode=normal" set of failregex instead.

We could have set all filters with a `[mode=%(mode)]` suffix, but there
really isn't a good point to do that. The mode used really does define
which set of failregex is used and thus should always be set to the
associated mode, otherwise we're not testing the appropriate thing.

While checking things up I've also seen that, now that the mode is
included in the common file, both Debian* hiera files don't add anything
else on top of the current values for common, so we can just get rid of
them.