Drop support for EOL OS releases: debian 10, ubuntu 18.04, RH 7, CentOS
In the most recent CI run, metadata2gha flagged all of those as
EOL, which they are, and automatically skipped running tests on them.

So we're in practice not testing those releases anymore: we might as
well stop claiming to be supporting them.

Note: this might not mean that this module will automatically stop
working on those releases.