Remove Vagrantfile and accompanying files
I'm moving towards a docker-based setup for local dev. The Vagrantfile
was useful but vagrant has no future wrt free software and it'll
probably get taken out of debian after a while. Also the tests directory
was a mess because of the hacks I made to get vagrant to somewhat work.