Bump version to 4.1.0
This release mainly adds support for ubuntu 24.04 and stdlib 9.

It also removes a couple of older OS releases which have become EOL.

Finally it adds a bunch of new parameters to complete the set of
available options.

The rest is just fiddling about with my local test setup and fixing the
tests in CI.